Move nvcc-only code from memory to backends::gpu (#342)
* Move gpu-kernel code from memory to backends/gpu

A small step towards seperate back end compilation for CUDA.
Move the following code to the gpu backend:
- the memory::fill* wrappers and fill kernel
- the managed_ptr type
  - only used in backends::gpu
  - has `__device__ __host__` members for dual host-device use.

* update unit&validation tests
