+++ src/CMakeLists.txt
@@ -27,7 +27,7 @@ set(BASE_SOURCES
 set(CUDA_SOURCES
     backends/gpu/fvm.cu
     backends/gpu/multi_event_stream.cu
-    memory/fill.cu
+    backends/gpu/fill.cu
 )
 
 # The cell_group_factory acts like an interface between the
